requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
DateTime==4.5
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
discord.py==1.7.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
Flask==2.2.2
fonttools==4.37.1
frozenlist==1.3.3
future==0.18.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpl-finance==0.10.1
multidict==6.0.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numexpr==2.8.3
numpy==1.21.6
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.10.0
pandastable==0.13.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1


requests==2.28.1
selenium==4.3.0
gunicorn==20.1.0
